Obtain SNAP climate data sets

Download climate data sets from SNAP.

climdata(id, area, time_scale = "monthly", set = NULL)

Arguments

id
character, data set ID. See collections for available data sets.

area
character, aerial unit/region or point location of interest. See climate_regions.

time_scale
character, "monthly", "seasonal" or "annual".

set
character or NULL, The set/group that area belongs to. Can be ignored, but should be provided in rare cases where area name is not unique (a warning is thrown). See climate_regions.

Value

a data frame.

Details

This function downloads climate data sets from Scenarios Network for Alaska and Arctic Planning based on the data set id and specification of the region or point location of interest.

Examples

climdata("ar5stats", "AK-CAN")

#> # A tibble: 101,400 x 20
#>           RCP   Model    Var  Group Region  Year  Month  Mean    SD   Min
#>        <fctr>  <fctr> <fctr>  <chr>  <chr> <int> <fctr> <dbl> <dbl> <dbl>
#>  1 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Jan  57.0  97.5   1.3
#>  2 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Feb  45.6  77.8   0.4
#>  3 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Mar  51.3  83.7   2.3
#>  4 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Apr  48.0  81.3   3.5
#>  5 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    May  51.5  47.6   0.2
#>  6 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Jun  80.7  52.9   5.7
#>  7 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Jul  71.1  53.5   0.8
#>  8 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Aug  65.4  73.7   0.0
#>  9 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Sep  80.5  79.0   4.3
#> 10 Historical CRU 4.0     pr AK-CAN AK-CAN  1901    Oct  73.7 132.2   2.1
#> # ... with 101,390 more rows, and 10 more variables: Max <dbl>, Pct_025 <dbl>,
#> #   Pct_05 <dbl>, Pct_10 <dbl>, Pct_25 <dbl>, Pct_50 <dbl>, Pct_75 <dbl>,
#> #   Pct_90 <dbl>, Pct_95 <dbl>, Pct_975 <dbl>